# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Node Tree API",
        "description": "API para la gestión de una estructura de árbol de nodos.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor de Desarrollo"
        }
    ],
    "paths": {
        "/api/nodes": {
            "post": {
                "tags": [
                    "Nodos"
                ],
                "summary": "Crear un nuevo nodo",
                "operationId": "cfafd706d8524f357d2d40d0c7747598",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "parent_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Nodo creado exitosamente."
                    },
                    "422": {
                        "description": "Error de validación."
                    }
                }
            }
        },
        "/api/nodes/roots": {
            "get": {
                "tags": [
                    "Nodos"
                ],
                "summary": "Listar todos los nodos raíz (sin parent_id)",
                "operationId": "b5c93f5e3426509a8e5c610ef1261c8e",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Elementos por página.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de nodos raíz paginada."
                    }
                }
            }
        },
        "/api/nodes/{nodeId}/children": {
            "get": {
                "tags": [
                    "Nodos"
                ],
                "summary": "Listar hijos y descendientes hasta la profundidad (depth) indicada",
                "operationId": "c8df4e31d45d44ffaf201fb9fa7f9e0a",
                "parameters": [
                    {
                        "name": "nodeId",
                        "in": "path",
                        "description": "ID del nodo padre.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "depth",
                        "in": "query",
                        "description": "Profundidad máxima de descendencia a buscar (1 para solo hijos directos).",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de descendientes paginada."
                    },
                    "404": {
                        "description": "Nodo padre no encontrado."
                    }
                }
            }
        },
        "/api/nodes/{nodeId}": {
            "delete": {
                "tags": [
                    "Nodos"
                ],
                "summary": "Elimina un nodo específico.",
                "description": "Solo se puede eliminar un nodo si NO tiene hijos.",
                "operationId": "deleteNode",
                "parameters": [
                    {
                        "name": "nodeId",
                        "in": "path",
                        "description": "ID del nodo a eliminar.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nodo eliminado exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nodo eliminado exitosamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nodo no encontrado."
                    },
                    "409": {
                        "description": "Error de conflicto de negocio (El nodo tiene hijos).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No se puede eliminar el nodo porque tiene hijos."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Nodos",
            "description": "Nodos"
        }
    ]
}